+++ microsite/src/theme/Footer.js
@@ -1,6 +1,15 @@
+import { Link } from '@material-ui/core'
 import React from 'react'
 import FooterImage from '../components/footerImage'
 import styles from './footer.module.css'
+
+function FooterLink({ link, text }) {
+  return (
+    <a className={styles.footerLicenseLink} target="_blank" href={link}>
+      {text}
+    </a>
+  )
+}
 
 function Footer() {
   return (
@@ -10,13 +19,16 @@ function Footer() {
         <p>
           Cloud Carbon Footprint is an open-source project, sponsored by
           ThoughtWorks Inc. under the&nbsp;
-          <a
-            className={styles.footerLicenseLink}
-            target="_blank"
-            href="http://www.apache.org/licenses/LICENSE-2.0"
-          >
-            Apache License, Version 2.0
-          </a>
+          <FooterLink
+            link="http://www.apache.org/licenses/LICENSE-2.0"
+            text="Apache License, Version 2.0"
+          />
+        </p>
+        <p>
+          <FooterLink
+            link="https://www.thoughtworks.com/privacy-policy"
+            text="PRIVACY POLICY"
+          />
         </p>
       </div>
     </div>