+++ packages/api/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @cloud-carbon-footprint/api
 
+## 0.1.1
+
+### Patch Changes
+
+- Updated dependencies [96cb023]
+  - @cloud-carbon-footprint/core@0.2.0
+
 ## 0.1.0
 
 ### Minor Changes