+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @cloud-carbon-footprint/cli
 
+## 0.2.0
+
+### Minor Changes
+
+- 96cb023: Adds support for Microsoft Azure, as well as updates to the microsite.
+
+### Patch Changes
+
+- Updated dependencies [96cb023]
+  - @cloud-carbon-footprint/core@0.2.0
+
 ## 0.1.0
 
 ### Minor Changes