+++ packages/client/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @cloud-carbon-footprint/client
 
+## 0.2.0
+
+### Minor Changes
+
+- 96cb023: Adds support for Microsoft Azure, as well as updates to the microsite.
+
 ## 0.1.0
 
 ### Minor Changes